(refactor) getCurrentData, getForecastData
-> getCurrentData now calls the class
CurrentWeather inside it, with all the info received
from the argument the function takes.
It only returns the object instead of all the variables
previously.

-> getForecastData calls the class DailyWeather
inside it instead of creating an object literal.

(refactor) locationWeather

-> gets all the data calls the necessary functions
creates a new class Location and then the objects
that come from getCurrentData, getForecastData and getHourlyData
are assigned to it with Object.assign.

(feat) getHourlyData

-> creates an array of new HourlyWeather classes, that
are returned to be assigned to locationWeather.